Unified server and local association manager implementation
